Changed amount to integer

+++ src/migrations/20260818_create_ledger_entries_table.js
@@ -6,7 +6,7 @@ export async function up() {
       id UUID PRIMARY KEY,
       user_id UUID NOT NULL REFERENCES users(id),
       currency VARCHAR(20) NOT NULL,
-      amount NUMERIC(12, 2) NOT NULL CHECK (amount <> 0),
+      amount INTEGER NOT NULL CHECK (amount <> 0),
       reason VARCHAR(50) NOT NULL,
       related_bet_id UUID REFERENCES bets(id),
       created_at TIMESTAMP NOT NULL DEFAULT CURRENT_TIMESTAMP
